+++ src/Yaat.Sim/Commands/ParsedCommand.cs
@@ -245,7 +245,7 @@ public record TurnDownwindCommand : ParsedCommand;
 
 public record TurnBaseCommand : ParsedCommand;
 
-public record ExtendDownwindCommand : ParsedCommand;
+public record ExtendPatternCommand : ParsedCommand;
 
 public record MakeShortApproachCommand : ParsedCommand;